Please WOrk

+++ src/main/java/com/sammy/malum/common/items/tools/ModAxeItem.java
@@ -6,7 +6,7 @@ import net.minecraftforge.common.ToolType;
 
 public class ModAxeItem extends AxeItem
 {
-    public ModAxeItem(IItemTier material, int damage, float speed, Properties properties)
+    public ModAxeItem(IItemTier material, float damage, float speed, Properties properties)
     {
         super(material, damage + 6, speed - 3.2f, properties.maxDamage(material.getMaxUses()).addToolType(ToolType.AXE, material.getHarvestLevel()));
     }
